separate logic and simulation code
